Python/Ejercicios/Prueba tecnica edunext/01_our_microservice/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/base.in requirements/dev.in requirements/test.in
#
appdirs==1.4.4
    # via virtualenv
asgiref==3.3.4
    # via django
astroid==2.5.6
    # via pylint
attrs==21.2.0
    # via pytest
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
click==8.0.1
    # via pip-tools
colorama==0.4.4
    # via radon
coverage==5.5
    # via pytest-cov
distlib==0.3.2
    # via virtualenv
django==3.2.3
    # via
    #   -r requirements/base.in
    #   djangorestframework
    #   jsonfield
djangorestframework==3.12.4
    # via -r requirements/base.in
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8-polyfill==1.0.2
    # via radon
flake8==3.9.2
    # via
    #   -r requirements/test.in
    #   flake8-polyfill
future==0.18.2
    # via radon
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via pylint
jsonfield==3.1.0
    # via -r requirements/base.in
lazy-object-proxy==1.6.0
    # via astroid
mando==0.6.4
    # via radon
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
packaging==20.9
    # via
    #   pytest
    #   tox
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements/dev.in
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.10.0
    # via
    #   pytest
    #   pytest-catchlog
    #   tox
pycodestyle==2.7.0
    # via
    #   -r requirements/test.in
    #   flake8
pyflakes==2.3.1
    # via flake8
pylint==2.8.3
    # via -r requirements/test.in
pyparsing==2.4.7
    # via packaging
pytest-catchlog==1.2.2
    # via -r requirements/test.in
pytest-cov==2.12.1
    # via -r requirements/test.in
pytest-django==4.3.0
    # via -r requirements/test.in
pytest==6.2.4
    # via
    #   pytest-catchlog
    #   pytest-cov
    #   pytest-django
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via xenon
radon==4.5.2
    # via
    #   -r requirements/test.in
    #   xenon
requests==2.25.1
    # via xenon
six==1.16.0
    # via
    #   mando
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via
    #   pep517
    #   pylint
    #   pytest
    #   pytest-cov
    #   tox
tox==3.23.1
    # via -r requirements/dev.in
urllib3==1.26.5
    # via requests
virtualenv==20.4.7
    # via tox
wrapt==1.12.1
    # via astroid
xenon==0.7.3
    # via -r requirements/test.in

# The following packages are considered to be unsafe in a requirements file:
# pip
